Add 'deleteScope' Method to Readme
I've noticed in the project source that there is a very useful `deleteScope` method for removing all events bound in a scope. 

I was struggling with keeping track of each bound key separately and found that this saves me a bit of time. 

Perhaps we can include in the docs so others can find this functionality more easily?